changed name of application
