+++ src/main/java/com/stucko09/statsaga/exception/UserDoesNotExistException.java
@@ -1,0 +1,11 @@
+package com.stucko09.statsaga.exception;
+
+public class UserDoesNotExistException extends RuntimeException {
+    public UserDoesNotExistException(String username) {
+        super("User with username " + username + " does not exist.");
+    }
+
+    public UserDoesNotExistException(long steamId) {
+        super("User with steamId " + steamId + " does not exist.");
+    }
+}
