+++ src/main/java/com/stucko09/statsaga/model/steam/SteamResponse.java
@@ -1,0 +1,8 @@
+package com.stucko09.statsaga.model.steam;
+
+import lombok.Data;
+
+@Data
+public class SteamResponse<T> {
+    private T response;
+}
