+++ src/main/java/com/stucko09/statsaga/util/Constants.java
@@ -1,0 +1,8 @@
+package com.stucko09.statsaga.util;
+
+public class Constants {
+
+    public enum PlaytimeRecordType {
+        INITIAL, DAILY
+    }
+}
