--- src/main/java/com/stucko09/steam_aggregator/SteamPlaytimeAggregatorApplication.java
@@ -1,11 +1,0 @@
-package com.stucko09.steam_aggregator;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SteamPlaytimeAggregatorApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(SteamPlaytimeAggregatorApplication.class, args);
-	}
-}
