--- src/main/java/com/stucko09/steam_aggregator/config/BasicAppConfig.java
@@ -1,21 +1,0 @@
-package com.stucko09.steam_aggregator.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-import com.github.dozermapper.core.DozerBeanMapperBuilder;
-import com.github.dozermapper.core.Mapper;
-
-@Configuration
-public class BasicAppConfig {
-    @Bean
-    RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-
-    @Bean
-    Mapper dozerBeanMapper() {
-        return DozerBeanMapperBuilder.buildDefault();
-    }
-}
