--- src/main/java/com/stucko09/steam_aggregator/config/SteamProperties.java
@@ -1,13 +1,0 @@
-package com.stucko09.steam_aggregator.config;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-import lombok.Data;
-
-@Data
-@Configuration
-@ConfigurationProperties(prefix = "steam")
-public class SteamProperties {
-    private String hostname;
-}
