--- src/main/java/com/stucko09/steam_aggregator/model/entity/BaseRecordClass.java
@@ -1,31 +1,0 @@
-package com.stucko09.steam_aggregator.model.entity;
-
-import java.time.ZonedDateTime;
-
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.MappedSuperclass;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@MappedSuperclass
-public abstract class BaseRecordClass {
-    @Setter(AccessLevel.NONE)
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Setter(AccessLevel.NONE)
-    @CreationTimestamp
-    private ZonedDateTime creationTimestamp;
-
-    @Setter(AccessLevel.NONE)
-    @UpdateTimestamp
-    private ZonedDateTime updateTimestamp;
-}
