--- src/main/java/com/stucko09/steam_aggregator/model/rest/GenericResponse.java
@@ -1,11 +1,0 @@
-package com.stucko09.steam_aggregator.model.rest;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class GenericResponse {
-    private String message;
-    private boolean success;
-}
