--- src/main/java/com/stucko09/steam_aggregator/model/steam/SteamResponse.java
@@ -1,8 +1,0 @@
-package com.stucko09.steam_aggregator.model.steam;
-
-import lombok.Data;
-
-@Data
-public class SteamResponse<T> {
-    private T response;
-}
