--- src/main/java/com/stucko09/steam_aggregator/util/Constants.java
@@ -1,8 +1,0 @@
-package com.stucko09.steam_aggregator.util;
-
-public class Constants {
-
-    public enum PlaytimeRecordType {
-        INITIAL, DAILY
-    }
-}
