--- src/test/java/com/stucko09/steam_aggregator/SteamPlaytimeAggregatorApplicationTests.java
@@ -1,14 +1,0 @@
-package com.stucko09.steam_aggregator;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@ActiveProfiles("junit")
-@SpringBootTest
-class SteamPlaytimeAggregatorApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-}
